requirements.txt
flask==2.0.1
pandas==1.3.3
numpy==1.21.2
torch==1.9.0
fair-esm==0.4.0
gunicorn==20.1.0 